+++ businesssim/src/com/FBLA/businesssim/entity/mob/Player.java
@@ -3,14 +3,14 @@ package com.FBLA.businesssim.entity.mob;
 import com.FBLA.businesssim.graphics.Screen;
 import com.FBLA.businesssim.graphics.Sprite;
 import com.FBLA.businesssim.input.Keyboard;
-import com.FBLA.businesssim.util.Vector2i;
+import com.FBLA.businesssim.util.Vector2d;
 
 public class Player extends Mob{
     
     private Screen screen;
     private Keyboard keys;
     
-    public Player(Vector2i v, Screen sc, Keyboard k)
+    public Player(Vector2d v, Screen sc, Keyboard k)
     {
         super(v,"Player");
         keys = k;
